- optimize defined npm test script
Removed the hard sleep statement of 30 seconds and configure wget call that determinate when the test server
instance is ready to try it 40 times and wait between this tries 10 seconds so that it wait for the server to be
ready 400 seconds in total. For a better continous local run of npm test first remove the RUNNING_PID file so that
play start will start a new server.

At the moment an play stop is missing for local run so if you made changes on the scala files you have to kill the
play server manualy and than you can perform npm test. A better solution will come in the future i hope but at the moment
it run very well on travis ci.