- implements new e2e test scenarion with a new user that has encrypted values store in the database
For the new test scenarion it was nessary to implements a possibility for the e2e tests
to specify which user the need for its tests. There are 2 users first user (email=testemail) with unencrypted values for balances and stocks
and second user (email=testemailenc) witn encrypted values for balances the default encryption key was used.

Backend changes: - added the new user (email=testemailenc)
		 - add play integration test that check if the new user exists

Frontend changes: - added new route in the app.js (/testlogin/:googleid) to login with different test users
		  - replaced hard coded testuser in oauthcontroller-mock.js with the :googleid route parameter

Added the new test scenario to the scenarios.js file.